SOFT: Extract SHA functions to mech_openssl.c
Extract the OpenSSL based SHA functions into file mech_openssl.c,
so that they can be used by other tokens as well.
